requirements.txt
flask==2.0.1
pymongo==4.3.3
marshmallow==3.19.0
python-dotenv==1.0.0
flask-cors==4.0.0
bcrypt==4.0.1
pydantic==2.0.2
flask-pydantic==0.11.0
python-jose==3.3.0
flask-smorest==0.42.0